Don't assume a device passed in from client
Note, that this will now only work as-is on systems with a single  wireless device.

To regain multi-device compatibility we would need to parse the JSON from the filesystem and look up the device name there.

Since we currently no longer have any development boards with multiple wireless interfaces, we leave this as-is.