Let tox handle the development requirements
This way we no longer need a separately maintained development virtualenv (tox now
builds that for us, so we still have one, if needed) and the tox environments will
contain all the dependencies required to test and build documentation.

Additionally, the `tox.ini` file is now the canonical place to define any and all
development and testing requirements and `setup.py` itself now only contains a
single reference to `tox`.