Build the frontend via travis
Note, that this requires installing a newer version of nodejs into the
travis environment, because travis only supports a single language per
project.

In the end it is easier to install newer version of nodejs into the
python image than the other way around, so that is what we're doing
here.