Make wifi onboarding more robust (#179)
* Update react-native-ble-plx to version 0.6.0

* Clean code

* Unsubscribe from hub notifications

* Display activity indicator while sending password

* Disable continue button when wifi isn’t connected

* Increase time interval for ignoring “disconnected”

* Disconnect from hub’s bluetooth adapter

* Fix hub sometimes not properly connected

We must not check for `isConnected()` when requesting a connection with `connect()`. The hub might be connected to the user’s phone via BLE, but we might not have subscribed for notifications and not yet read some characteristics. Furthermore one single catch-block needs to go the very end as before these changes, failed `connect()` attempts will still result in subscription requests and read requests.

* Handle hub connection failure

* Clean code

* Render wifi connection state

* Stop sending connection state notifications

* Add todo comment

* Make code pep8 conform

* Clean code

* Add todo comment