Make the on-boarding process work on iOS9 (#282)
Closes #280

Allow setup_app to connect to the hub via a local network without HTTPS.
Previously it worked only on iOS10. Read more about it here:

  https://dev.tapjoy.com/faq/app-transport-security-ats-ios-10/

This is only a workaround for development, we'll have to decide on how to make
it more secure for production devices. For this I opened #281.