Restarting BT on unsuccessful nuimo connection
The previous implementation addressed unsuccessful connects
by reconnect attempts. That didn't solve the problem for the
new module which in case it didn't connect on the first try
it would never connect.

The implementation adds hard physical and os wise resets of
the bt subsytems once a nuimo cand be detected or connected to.

Timeouts and number of retries are defined as globals on the top.

This implementation doesn't take into consideration the onboarding
of multiple nuimos.